Program structure changes.
Moved Person class to own file and added dependency to pizzaCalc
