+++ pizzacalc.py
@@ -1,26 +1,4 @@
-class Person(object):
-
-    moneyOwed = 0
-
-    def __init__(self):
-        self.name = self.getName()
-
-    def getName(self):
-        return input("Name?: ")
-        
-    def getcost(self):
-        while(True):
-            try:
-                cost = float(input("Cost?: "))
-                break
-            except ValueError:
-                print ("Please input a number: ")
-        return cost
-
-    def addPizza(self):
-        self.moneyOwed += self.getcost()
-        return self.moneyOwed
-
+from Person import Person
 
 def calcTotalCost(people, tax):
     totalCost = 0
